Fix base sensor tests.
